requirements.txt
feedparser
requests
beautifulsoup4
python-telegram-bot